+++ sdk/config.h
@@ -104,7 +104,7 @@ public:
   // Post to Call Ratio
   float post_to_call_ratio() const { return config_.post_call_ratio; }
   // Max number of emails waiting allowed
-  uint8_t max_waiting() const { return config_.maxwaiting; }
+  int max_waiting() const { return static_cast<int>(config_.maxwaiting); }
   // Directory number where uploads go by default.
   uint8_t new_uploads_dir() const { return config_.newuploads; }
